refactor: rename iterators to events
