+++ src/graphql/resolvers/mutations/update-todo.js
@@ -1,6 +1,6 @@
 import { returnTodo } from '../../../database/utils/utils'
 import { pubsub } from '../../../server/pubsub'
-import { TODO_UPDATED } from '../iterators'
+import { TODO_UPDATED } from '../events'
 
 export const updateTodo = async (
   root,
