+++ src/graphql/resolvers/subscriptions/todo-added.js
@@ -1,6 +1,6 @@
 import { withFilter } from 'apollo-server'
 import { pubsub } from '../../../server/pubsub'
-import { TODO_ADDED } from '../iterators'
+import { TODO_ADDED } from '../events'
 
 export default {
   resolve: (payload) => payload.todoAdded,
