+++ src/graphql/resolvers/subscriptions/todo-status-updated.js
@@ -1,6 +1,6 @@
 import { withFilter } from 'apollo-server'
 import { pubsub } from '../../../server/pubsub'
-import { TODO_STATUS_UPDATED } from '../iterators'
+import { TODO_STATUS_UPDATED } from '../events'
 
 export default {
   resolve: (payload) => payload.todoStatusUpdated,
